+++ src/main/java/com/x2ge/mqtt/MessageData.java
@@ -1,4 +1,4 @@
-package com.wge.netty.mqtt;
+package com.x2ge.mqtt;
 
 import java.io.Serializable;
