+++ src/main/java/com/x2ge/mqtt/ProcessorResult.java
@@ -1,4 +1,4 @@
-package com.wge.netty.mqtt;
+package com.x2ge.mqtt;
 
 public class ProcessorResult {
