+++ src/main/java/com/x2ge/mqtt/utils/Log.java
@@ -1,4 +1,4 @@
-package com.wge.netty.mqtt.utils;
+package com.x2ge.mqtt.utils;
 
 import java.lang.management.ManagementFactory;
 import java.text.SimpleDateFormat;
